feat(post): adicionar data de criação automática
Adiciona o campo dataCriacao à entidade Post, que armazena automaticamente a data e hora da criação do post.

- O campo é do tipo LocalDateTime no Java e mapeado como TIMESTAMP no banco de dados.
- Utiliza a anotação @PrePersist para definir a data antes da inserção no banco.
- Garante que o valor seja preenchido corretamente ao salvar um novo post.

Essa mudança permite rastrear quando um post foi criado sem necessidade de preenchimento manual.

+++ src/main/java/com/example/overclockAPI/services/UsuarioService.java
@@ -34,7 +34,12 @@ public class UsuarioService {
     }
 
     public Usuario deletarUsuario(Long id) {
-        usuarioRepository.deleteById(id);
-        return usuarioRepository.findById(id).get();
+        if (usuarioRepository.existsById(id)){
+            usuarioRepository.deleteById(id);
+        }
+        else{
+            System.out.print("Esse usuario não existe");
+        }
+        return null;
     }
 }
